requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
fastapi
uvicorn
anthropic
google-generativeai